// src/data/generatedData.ts
export const mapCodes: MapCode[] = [
  {
    "code": "550",
    "name": "滁州市",
    "type": "latn_id"
  },
  {
    "code": "551",
    "name": "合肥市",
    "type": "latn_id"
  },
  {
    "code": "552",
    "name": "蚌埠市",
    "type": "latn_id"
  },
  {
    "code": "553",
    "name": "芜湖市",
    "type": "latn_id"
  },
  {
    "code": "554",
    "name": "淮南市",
    "type": "latn_id"
  },
  {
    "code": "555",
    "name": "马鞍山市",
    "type": "latn_id"
  },
  {
    "code": "556",
    "name": "安庆市",
    "type": "latn_id"
  },
  {
    "code": "557",
    "name": "宿州市",
    "type": "latn_id"
  },
  {
    "code": "558",
    "name": "阜阳市",
    "type": "latn_id"
  },
  {
    "code": "559",
    "name": "黄山市",
    "type": "latn_id"
  },
  {
    "code": "560",
    "name": "亳州市",
    "type": "latn_id"
  },
  {
    "code": "561",
    "name": "淮北市",
    "type": "latn_id"
  },
  {
    "code": "562",
    "name": "铜陵市",
    "type": "latn_id"
  },
  {
    "code": "563",
    "name": "宣城市",
    "type": "latn_id"
  },
  {
    "code": "564",
    "name": "六安市",
    "type": "latn_id"
  },
  {
    "code": "566",
    "name": "池州市",
    "type": "latn_id"
  },
  {
    "code": "100000",
    "name": "暂存",
    "type": "status_cd"
  },
  {
    "code": "100004",
    "name": "已作废",
    "type": "status_cd"
  },
  {
    "code": "101100",
    "name": "待审核",
    "type": "status_cd"
  },
  {
    "code": "101200",
    "name": "待收费",
    "type": "status_cd"
  },
  {
    "code": "101300",
    "name": "待确认",
    "type": "status_cd"
  },
  {
    "code": "101400",
    "name": "客户需求确认",
    "type": "status_cd"
  },
  {
    "code": "201298",
    "name": "缓装中",
    "type": "status_cd"
  },
  {
    "code": "201300",
    "name": "开通中",
    "type": "status_cd"
  },
  {
    "code": "201400",
    "name": "已开通",
    "type": "status_cd"
  },
  {
    "code": "201700",
    "name": "待发送",
    "type": "status_cd"
  },
  {
    "code": "300000",
    "name": "竣工",
    "type": "status_cd"
  },
  {
    "code": "301100",
    "name": "竣工中",
    "type": "status_cd"
  },
  {
    "code": "401300",
    "name": "已撤单",
    "type": "status_cd"
  },
  {
    "code": "401397",
    "name": "删单",
    "type": "status_cd"
  },
  {
    "code": "401400",
    "name": "已退单",
    "type": "status_cd"
  }
]